requirements.txt
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
dnspython==2.4.2
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-vision==3.4.4
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
idna==3.4
proto-plus==1.22.3
protobuf==4.24.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pymongo==4.4.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
schedule==1.2.0
six==1.16.0
soupsieve==2.4.1
urllib3==1.26.16
